feat: fully compat old APIs
