+++ api_message_test.go
@@ -22,6 +22,11 @@ func TestPostText(t *testing.T) {
 		assert.Equal(t, 0, resp.Code)
 		assert.NotEmpty(t, resp.Data.MessageID)
 	}
+	resp, err = bot.PostText("PostText: union_id hello, world", WithUnionID(testUserUnionID))
+	if assert.NoError(t, err) {
+		assert.Equal(t, 0, resp.Code)
+		assert.NotEmpty(t, resp.Data.MessageID)
+	}
 }
 
 func TestPostTextFailed(t *testing.T) {
@@ -67,6 +72,14 @@ func TestPostShareChat(t *testing.T) {
 	}
 }
 
+func TestPostShareUser(t *testing.T) {
+	resp, err := bot.PostShareUser(testUserOpenID, WithChatID(testGroupChatID))
+	if assert.NoError(t, err) {
+		assert.Equal(t, 0, resp.Code)
+		assert.NotEmpty(t, resp.Data.MessageID)
+	}
+}
+
 func TestPostMessage(t *testing.T) {
 	// text message
 	msg := NewMsgBuffer(MsgText)
@@ -100,6 +113,14 @@ func TestPostMessage(t *testing.T) {
 		assert.Equal(t, 0, resp.Code)
 		assert.NotEmpty(t, resp.Data.MessageID)
 	}
+	// share user
+	msg = NewMsgBuffer(MsgShareUser)
+	om = msg.BindChatID(testGroupChatID).ShareUser(testUserOpenID).Build()
+	resp, err = bot.PostMessage(om)
+	if assert.NoError(t, err) {
+		assert.Equal(t, 0, resp.Code)
+		assert.NotEmpty(t, resp.Data.MessageID)
+	}
 }
 
 func TestPostPostMessage(t *testing.T) {
@@ -124,62 +145,6 @@ func TestPostPostMessage(t *testing.T) {
 }
 
 func TestPostCardMessage(t *testing.T) {
-	cardContentV4 := `{
-		"config": {
-			"wide_screen_mode": false
-		},
-		"elements": [
-			{
-				"tag": "div",
-				"text": {
-					"i18n": {
-						"zh_cn": "中文文本",
-						"en_us": "English text",
-						"ja_jp": "日本語文案"
-					},
-					"tag": "plain_text"
-				}
-			},
-			{
-				"tag": "div",
-				"text": {
-					"tag": "plain_text",
-					"content": "This is a very very very very very very very long text;"
-				}
-			},
-			{
-				"actions": [
-					{
-						"tag": "button",
-						"text": {
-							"content": "a",
-							"tag": "plain_text"
-						},
-						"type": "default"
-					}
-				],
-				"tag": "action"
-			}
-		],
-		"header": {
-			"title": {
-				"content": "a",
-				"tag": "plain_text"
-			}
-		}
-	}
-	`
-
-	msgV4 := NewMsgBuffer(MsgInteractive)
-	omV4 := msgV4.BindEmail(testUserEmail).Card(cardContentV4).Build()
-	resp, err := bot.PostMessage(omV4)
-	if assert.NoError(t, err) {
-		assert.Equal(t, 0, resp.Code)
-		assert.NotEmpty(t, resp.Data.MessageID)
-	}
-}
-
-func TestPostCardBuilderMessage(t *testing.T) {
 	b := NewCardBuilder()
 	card := b.Card(
 		b.Div(
@@ -222,23 +187,7 @@ func TestPostCardBuilderMessage(t *testing.T) {
 	}
 }
 
-func TestPostRichText(t *testing.T) {
-	postContent := NewPostBuilder().
-		Title("post title").
-		TextTag("hello, world", 1, true).
-		LinkTag("Google", "https://google.com/").
-		AtTag("www", testGroupChatID).
-		ImageTag("img_a7c6aa35-382a-48ad-839d-d0182a69b4dg", 300, 300).
-		Render()
-	resp, err := bot.PostRichText(postContent, WithEmail(testUserEmail))
-	t.Log(resp)
-	if assert.NoError(t, err) {
-		assert.Equal(t, 0, resp.Code)
-		assert.NotEmpty(t, resp.Data.MessageID)
-	}
-}
-
-func TestPostEphemeralMessage(t *testing.T) {
+func TestEphemeralMessage(t *testing.T) {
 	b := NewCardBuilder()
 	card := b.Card(
 		b.Div().
@@ -260,29 +209,33 @@ func TestPostEphemeralMessage(t *testing.T) {
 	}
 }
 
-func TestRecallMessage(t *testing.T) {
-	resp, err := bot.PostText("PostText: open_chat_id hello, world", WithChatID(testGroupChatID))
+func TestMessageCRUD(t *testing.T) {
+	// create
+	resp, err := bot.PostText("PostText: CRUD hello, world", WithChatID(testGroupChatID))
 	if assert.NoError(t, err) {
 		assert.Equal(t, 0, resp.Code)
 		assert.NotEmpty(t, resp.Data.MessageID)
+		t.Log(resp.Data)
 	}
+	// get
+	getResp, err := bot.GetMessage(resp.Data.MessageID)
+	if assert.NoError(t, err) {
+		assert.Equal(t, 0, getResp.Code)
+		if assert.NotEmpty(t, getResp.Data.Items) {
+			assert.Equal(t, getResp.Data.Items[0].MessageID, resp.Data.MessageID)
+		}
+	}
+	// recall
 	rcResp, err := bot.RecallMessage(resp.Data.MessageID)
 	if assert.NoError(t, err) {
 		assert.Equal(t, 0, rcResp.Code)
 	}
-}
-
-func TestMessageReceipt(t *testing.T) {
-	resp, err := bot.PostText("Message that needs receipt", WithChatID(testGroupChatID))
-	if assert.NoError(t, err) {
-		assert.Equal(t, 0, resp.Code)
-		assert.NotEmpty(t, resp.Data.MessageID)
-	}
+	// receipt unread
 	receipt, err := bot.MessageReadReceipt(resp.Data.MessageID)
 	if assert.NoError(t, err) {
 		t.Log(receipt.Data.ReadUsers)
 	}
-
+	// receipt read
 	receiptOld, err := bot.MessageReadReceipt(testMessageID)
 	if assert.NoError(t, err) {
 		assert.NotEmpty(t, receiptOld.Data.ReadUsers)
